[WIP] Basic GTP/0.1 PWP Client, with handshake
Enough code to start a basic PWP node up with a known list of peers,
connect to them and blow a raspberry^W send a choke message, and shut
down.
    
A couple of trivial changes were pulled from the 'core' branch, and
this does indicate that there should be a new module that represents
an association - connection - between two peers.

There are some other unrelated changes in this patch which need to
move onto the "core" branch.