Replace some confusing Coro code in the PWP handshake test
While setting up the two-node handshake test, as each node is running
in a separate Coro it is required that we let control pass from one
thread to the other during the test.  The way that this was previously
done was to just delay until the state was observed to change, but
this is not required - the message has been sent on the socket, and
the problem was that Coro was just not checking the socket for
messages with a simple "cede", as it had already been checked in the
last "sweep" of filehandles.

So, make t/50-pwp-handshake.t use Coro::Event::loop with a timeout
when the timeout is really required (on initial connection), and use
Coro::Event::sweep to mark the background threads as unchecked so that
the test can proceed directly.