Partial implementation of the 'Commit Reel'
We start with a rev-list of all the commits in topographical order.  Then,
each commit has the objects it depends on queried, again using rev-list.
The offset in the reel, object type, object size, and object hash are
printed to stdout.

Right now we do not take into account commit time or SHA1 hash for ordering
purposes, according the the RFC.