Fix sending and receiving References
Right now we keep lists of References in 3 places, which may or may not be
necessary.  We have a global list in Torrent, and a list of References sent
by and from each peer.  This should help us follow the RFC when it says not
to send out References to a peer more than once per connection, though.